// src/Commands.d.ts
export declare const Commands: {
    "info": {
        "basicStatus": string;
        "playerInfo": string;
        "selectedInput": string;
        "systemInfo": string;
    };
    "input": {
        "airplay": string;
        "aux1": string;
        "aux2": string;
        "cd": string;
        "digital1": string;
        "digital2": string;
        "netRadio": string;
        "server": string;
        "spotify": string;
        "tuner": string;
        "usb": string;
    };
    "playback": {
        "pause": string;
        "play": string;
        "rewind": string;
        "skip": string;
        "stop": string;
    };
    "power": {
        "off": string;
        "on": string;
    };
    "sound": {
        "muteOff": string;
        "muteOn": string;
        "volume": string;
    };
};